xnumx将结果储存为double类型

+++ project/USER/inc/xnumx.h
@@ -3,7 +3,7 @@
 
 #include "stm32f10x.h"
 
-void xnumx(char * rx,char rxLen,char * separator,float * mwData,char mwDataLen);
+void xnumx(char * rx,char rxLen,char * separator,double * mwData,char mwDataLen);
